Fix and extend equal class sampling to sklearn code.

+++ cifar10_cnn.py
@@ -32,7 +32,7 @@ n_train=1000
 classes=np.unique(y_train)
 inds_all=np.array([],dtype='int32')
 for cl in classes:
-    inds=np.random.choice(np.where(np.array(y_train)==cl)[0], n_train/len(classes))
+    inds=np.random.choice(np.where(np.array(y_train)==cl)[0], int(n_train/len(classes)))
     inds_all=np.r_[inds,inds_all]
     
 x_train = x_train[inds_all]
